Remove dataset outputter from Benchmark class
Move to `moveit_benchmark_suite_output` pkg.
