+++ core/include/moveit_benchmark_suite/benchmark.h
@@ -46,41 +46,6 @@
 #include <moveit_benchmark_suite/profiler.h>
 
 namespace moveit_benchmark_suite {
-/** \brief An abstract class for outputting benchmark results.
- */
-class DataSetOutputter
-{
-public:
-  /** \brief Virtual destructor for cleaning up resources.
-   */
-  virtual ~DataSetOutputter() = default;
-
-  /** \brief Write the \a results of a benchmarking query out.
-   *  Must be implemented by child classes.
-   *  \param[in] results The results of one query of benchmarking.
-   */
-  virtual void dump(const DataSet& results, const std::string& pathname) = 0;
-};
-
-class BenchmarkSuiteDataSetOutputter : public DataSetOutputter
-{
-public:
-  /** \brief Constructor.
-   *  \param[in] prefix Prefix to place in front of all log files generated.
-   *  \param[in] dumpScene If true, will output scene into log file.
-   */
-  BenchmarkSuiteDataSetOutputter();
-
-  /** \brief Destructor, runs `ompl_benchmark_statistics.py` to generate benchmarking database.
-   */
-  ~BenchmarkSuiteDataSetOutputter() override;
-
-  /** \brief Dumps \a results into a OMPL benchmarking log file in \a prefix_ named after the request \a
-   *  name_.
-   *  \param[in] results Results to dump to file.
-   */
-  void dump(const DataSet& dataset, const std::string& pathname) override;
-};
 
 class Benchmark
 {
@@ -153,7 +118,5 @@ private:
   std::vector<PostTrialCallback> post_trial_callbacks_;
   std::vector<PostQueryCallback> post_query_callbacks_;
   std::vector<PostBenchmarkCallback> post_benchmark_callbacks_;
-
-  BenchmarkSuiteDataSetOutputter outputter_;
 };
 }  // namespace moveit_benchmark_suite
