+++ core/src/benchmark.cpp
@@ -10,11 +10,6 @@ bool Benchmark::initialize(const std::string& name, const Options& options)
   name_ = name;
   options_ = options;
 
-  // Output dataset to logfile
-  {
-    addPostBenchmarkCallback([=](DataSetPtr& dataset) { outputter_.dump(*dataset, options_.output_file); });
-  }
-
   return true;
 };
 
@@ -208,73 +203,4 @@ void Benchmark::fillMetaData(DataSetPtr& dataset) const
   // dataset->metadata[DATASET_DATE_KEY] = to_simple_string(dataset->date);
   // dataset->metadata[DATASET_TOTAL_TIME_KEY] = dataset->time;
   // dataset->metadata[DATASET_CONFIG_KEY] = dataset->query_setup.query_setup;
-}
-
-///
-/// BenchmarkSuiteDataSetOutputter
-///
-
-BenchmarkSuiteDataSetOutputter::BenchmarkSuiteDataSetOutputter()
-{
-}
-
-BenchmarkSuiteDataSetOutputter::~BenchmarkSuiteDataSetOutputter()
-{
-}
-
-void BenchmarkSuiteDataSetOutputter::dump(const DataSet& dataset, const std::string& pathname)
-{
-  std::ofstream out;
-  std::string out_file;
-  std::string out_filepath;
-  std::string out_filename;
-
-  auto filepath = IO::getFilePath(pathname);
-  auto filename = IO::getFileName(pathname);
-
-  // Create filename if not specified and add extension
-  out_filename = filename;
-  if (out_filename.empty())
-    out_filename = log::format("%1%_%2%", dataset.name, IO::getDateStr() + ".yaml");
-
-  // Set filepath as ROS_HOME
-  out_filepath = filepath;
-  if (out_filepath.empty())
-    out_filepath = IO::getEnvironmentPath("ROS_HOME");
-
-  // Set filepath as default ROS default home path
-  if (out_filepath.empty())
-  {
-    out_filepath = IO::getEnvironmentPath("HOME");
-    out_filepath = out_filepath + "/.ros";
-  }
-  else if (out_filepath[0] != '/')
-  {
-    std::string tmp = out_filepath;
-    out_filepath = IO::getEnvironmentPath("HOME");
-    out_filepath = out_filepath + "/.ros";
-    out_filepath = out_filepath + "/" + tmp;
-  }
-
-  if (!out_filepath.empty() && out_filepath.back() != '/')
-    out_filepath = out_filepath + '/';
-
-  if (!IO::createFile(out, out_filepath + out_filename))
-  {
-    ROS_ERROR_STREAM(log::format("File creation failed for: '%1%'", out_filepath + out_filename));
-    return;
-  }
-
-  YAML::Node node;
-  node["dataset"] = dataset;
-
-  // Add dataset as a sequence
-  YAML::Node root_node;
-  root_node.push_back(node);
-
-  out << "\n";  // Necessary for appending a dataset with right indentation
-  out << root_node;
-  out.close();
-
-  ROS_INFO_STREAM(log::format("Successfully created dataset: '%1%'", out_filepath + out_filename));
 }
